// src/api/models/User.ts
export type User = {
    createTime?: string;
    gender?: number | null;
    id?: number | null;
    isDelete?: number | null;
    isVip?: number | null;
    updateTime?: string;
    userAccount?: string;
    userAvatar?: string;
    userName?: string;
    userPassword?: string;
    userRole?: string;
};